Implementing updated keeper package
