+++ package.json
@@ -15,10 +15,9 @@
   "private": true,
   "scripts": {
     "postinstall": "eth-sdk",
-    "start:propagate": "ts-node src/propagate-job.ts",
+    "start:propagate": "ts-node src/init-propagate.ts",
     "build": "eth-sdk tsc",
-    "prod:order": "node dist/src/order-job.js",
-    "prod:deposit-manager": "node dist/src/deposit-manager-job.js",
+    "prod:propagate": "node dist/src/init-propagate.js",
     "lint": "xo",
     "lint:fix": "xo --fix"
   },
@@ -30,7 +29,7 @@
     "@connext/nxtp-adapters-subgraph": "^0.2.8-alpha.1",
     "@connext/nxtp-utils": "^0.2.8-alpha.1",
     "@eth-optimism/sdk": "^2.0.0",
-    "@keep3r-network/keeper-scripting-utils": "0.0.0-d9b60cef",
+    "@keep3r-network/keeper-scripting-utils": "0.0.0-554bcc1d",
     "dotenv": "16.0.3",
     "ethers": "5.6.9",
     "lodash.isequal": "4.5.0",
