--- src/utils/contants.ts
@@ -1,24 +1,0 @@
-/// COMMON CONSTANTS ///
-
-// Ethereum Mainnet
-export const CHAIN_ID = 1;
-
-// Size of our batch of bundles
-export const BURST_SIZE = 3;
-
-// Blocks into the future to send our first batch of bundles
-export const FUTURE_BLOCKS = 1;
-
-// Priority fee to use
-export const PRIORITY_FEE = 2;
-
-// Flashbots Relayer URL
-export const FLASHBOTS_RPC = 'https://relay.flashbots.net';
-
-/// JOBS CONSTANTS ///
-
-// Orderer API URL
-export const ORDER_API_URL = 'https://api.phuture.finance/orders?chain_id=1';
-
-// USDC Savings Vault Contract Address
-export const USV_ADDRESS = '0x6bAD6A9BcFdA3fd60Da6834aCe5F93B8cFed9598';
